Failing test -- commiting to save -- need to check further

+++ src/backend/api/src/main/java/dev/agasen/ecom/api/core/order/event/OrderEvent.java
@@ -15,7 +15,7 @@ public interface OrderEvent extends DomainEvent, OrderSaga {
                  Long customerId,
                  int quantity,
                  int unitPrice,
-                 Long totalAmount) implements OrderEvent {}
+                 int totalAmount) implements OrderEvent {}
 
   @Builder
   record Cancelled(Long orderId,
