+++ src/backend/api/src/main/java/dev/agasen/ecom/api/core/payment/model/PaymentProcessRequest.java
@@ -3,6 +3,6 @@ package dev.agasen.ecom.api.core.payment.model;
 import lombok.Builder;
 
 @Builder
-public record PaymentProcessRequest(Long customerId, Long orderId, Long amount) {
+public record PaymentProcessRequest(Long customerId, Long orderId, int amount) {
 
 }
